DisplaySettings: convert animation FPS from Integer to Double.
This fixes issue #123.

+++ mmstudio/src/main/java/org/micromanager/display/internal/ScrollerPanel.java
@@ -138,7 +138,7 @@ public class ScrollerPanel extends JPanel {
 
    private Timer snapbackTimer_;
    private Timer animationTimer_;
-   private int animationFPS_ = 0;
+   private double animationFPS_ = 0.0;
    private int animationStepSize_ = 0;
    private long lastAnimationTimeMs_ = 0;
 
@@ -167,10 +167,10 @@ public class ScrollerPanel extends JPanel {
 
       // Set up the FPS rate prior to calling addScroller(), below, as
       // that method creates the FPS button which needs animationFPS_ to be set
-      Integer fps = display.getDisplaySettings().getAnimationFPS();
+      Double fps = display.getDisplaySettings().getAnimationFPS();
       // Default to 10 if it's not set.
       if (fps == null) {
-         fps = 10;
+         fps = 10.0;
          display.setDisplaySettings(display.getDisplaySettings().copy()
                .animationFPS(fps).build());
       }
@@ -566,7 +566,7 @@ public class ScrollerPanel extends JPanel {
       // by for each animation tick. We cap at an update rate of 30FPS;
       // past that, we step forward by more than one image per step to
       // compensate.
-      int updateFPS = Math.min(30, animationFPS_);
+      double updateFPS = Math.min(30.0, animationFPS_);
       animationStepSize_ = (int) Math.max(1,
             Math.round(updateFPS * animationFPS_ / 1000.0));
 
@@ -597,7 +597,7 @@ public class ScrollerPanel extends JPanel {
       }
       else {
          // Target delay between updates
-         int delayMs = 1000 / updateFPS;
+         int delayMs = (int) (1000 / updateFPS);
          long sleepTime = Math.max(0,
                delayMs - (System.currentTimeMillis() - lastAnimationTimeMs_));
          animationTimer_.schedule(task, sleepTime);
